pyproject.toml
[project]
name = "neuro-mirror"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "anthropic>=0.77.0",
    "google-generativeai>=0.8.6",
    "gradio>=4.0.0",
    "gtts>=2.5.4",
    "huggingface-hub>=1.3.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "mediapipe==0.10.14",
    "numpy>=2.2.6",
    "openai>=2.16.0",
    "opencv-python>=4.13.0.90",
    "plotly>=6.5.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "reachy-mini[mujoco]>=1.2.13",
    "scipy>=1.15.3",
    "textblob>=0.19.0",
    "webrtcvad>=2.0.10",
]
